[refactor] move SeekHandler out of utils into root
