+++ xbmc/SeekHandler.cpp
