+++ xbmc/SeekHandler.h
