+++ xbmc/interfaces/builtins/PlayerBuiltins.cpp
@@ -28,6 +28,7 @@
 #include "GUIUserMessages.h"
 #include "PartyModeManager.h"
 #include "PlayListPlayer.h"
+#include "SeekHandler.h"
 #include "settings/AdvancedSettings.h"
 #include "settings/MediaSettings.h"
 #include "settings/Settings.h"
@@ -35,7 +36,6 @@
 #include "system.h"
 #include "utils/FileExtensionProvider.h"
 #include "utils/log.h"
-#include "utils/SeekHandler.h"
 #include "utils/StringUtils.h"
 #include "utils/URIUtils.h"
 #include "view/GUIViewState.h"
